requirements.txt
altair==6.0.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
cachetools==7.0.5
captcha==0.7.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dash==4.1.0
dash-bootstrap-components==2.0.4
extra-streamlit-components==0.1.81
Flask==3.1.3
Flask-Caching==2.3.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
pluggy==1.6.0
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
retrying==1.4.2
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
streamlit==1.55.0
streamlit-authenticator==0.4.2
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.8
wheel==0.46.3
zipp==3.23.0
